Fix bug introduced in f502bdf4903545cfc75dfe15bed650138da53a81

+++ libyara/include/yara/arena.h
@@ -31,7 +31,7 @@ limitations under the License.
 
 typedef struct _YR_RELOC
 {
-  size_t offset;
+  uint32_t offset;
   struct _YR_RELOC* next;
 
 } YR_RELOC;
